refactor(frontend): solved conflicts
