+++ packages/frontend/components/Shared/AddTokenButton.tsx
@@ -1,8 +1,8 @@
 import LoadingButton from '@mui/lab/LoadingButton';
 import { useState } from 'react';
 
+import { SerializableToken } from '../../helpers/history';
 import { addTokenToMetamask } from '../../helpers/metamask';
-import { SerializableToken } from '../../store/history.store';
 
 type ButtonAddTokenProps = {
   token: SerializableToken;
